requirements.txt
aiodns==3.0.0
aiosmtpd==1.4.2
arabic-reshaper==2.1.3
asgiref==3.3.4
atpublic==2.3
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
crispy==0.6.3
cycler==0.10.0
Django==3.2.3
django-countries==7.2.1
django-country-dialcode==0.6.3
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
django-jazzmin==2.4.7
django-multiselectfield==0.1.12
django-tinymce==3.3.0
django-widget-tweaks==1.4.8
fabio==0.11.0
fpdf==1.7.2
future==0.18.2
h5py==3.3.0
heroku==0.1.4
html5lib==1.1
idna==3.2
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.2
mysqlclient==2.0.3
numpy==1.21.0
pandas==1.2.5
Pillow==8.2.0
psycopg2==2.8.6
pycares==4.0.0
pycparser==2.20
pyparsing==2.4.7
PyPDF2==1.26.0
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.1
reportlab==3.6.2
requests==2.26.0
scikit-learn==0.24.2
scipy==1.7.1
silx==0.15.2
six==1.16.0
sqlparse==0.4.1
threadpoolctl==2.2.0
urllib3==1.26.6
validate-email==1.3
verify-email==2.4.3
waitress==2.0.0
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
